+++ lib/packages/wikibase/data-model/src/Claim/Claims.php
@@ -22,7 +22,7 @@ use Wikibase\DataModel\Statement\Statement;
  * @author Daniel Kinzler
  * @author H. Snater < mediawiki@snater.com >
  */
-class Claims extends ArrayObject implements Hashable, Comparable {
+class Claims extends ArrayObject implements Hashable {
 
 	/**
 	 * @see GenericArrayObject::__construct
@@ -280,76 +280,6 @@ class Claims extends ArrayObject implements Hashable, Comparable {
 	}
 
 	/**
-	 * Get array of Claim guids
-	 *
-	 * @since 0.4
-	 *
-	 * @return string[]
-	 */
-	public function getGuids() {
-		return array_map( function ( Claim $claim ) {
-			return $claim->getGuid();
-		}, iterator_to_array( $this ) );
-	}
-
-	/**
-	 * Returns the claims for the given property.
-	 *
-	 * @since 0.4
-	 *
-	 * @param PropertyId $propertyId
-	 *
-	 * @return Claims
-	 */
-	public function getClaimsForProperty( PropertyId $propertyId ) {
-		$byPropertyIdGrouper = new ByPropertyIdGrouper( $this );
-
-		if ( !$byPropertyIdGrouper->hasPropertyId( $propertyId ) ) {
-			return new self();
-		}
-
-		return new self( $byPropertyIdGrouper->getByPropertyId( $propertyId ) );
-	}
-
-	/**
-	 * Returns the main Snaks of the claims in this list.
-	 *
-	 * @since 0.4
-	 *
-	 * @return Snak[]
-	 */
-	public function getMainSnaks() {
-		$snaks = array();
-
-		/* @var Claim $claim */
-		foreach ( $this as $claim ) {
-			$guid = $claim->getGuid();
-			$snaks[$guid] = $claim->getMainSnak();
-		}
-
-		return $snaks;
-	}
-
-	/**
-	 * Returns a map from GUIDs to claim hashes.
-	 *
-	 * @since 0.4
-	 *
-	 * @return string[]
-	 */
-	public function getHashes() {
-		$snaks = array();
-
-		/* @var Claim $claim */
-		foreach ( $this as $claim ) {
-			$guid = $claim->getGuid();
-			$snaks[$guid] = $claim->getHash();
-		}
-
-		return $snaks;
-	}
-
-	/**
 	 * Returns a hash based on the value of the object.
 	 * The hash is based on the hashes of the claims contained,
 	 * with the order of claims considered significant.
@@ -374,109 +304,6 @@ class Claims extends ArrayObject implements Hashable, Comparable {
 	 */
 	public function isEmpty() {
 		return !$this->getIterator()->valid();
-	}
-
-	/**
-	 * Returns a new instance only containing the claims with the given rank.
-	 *
-	 * @since 0.7
-	 *
-	 * @param int $rank
-	 *
-	 * @return Claims
-	 */
-	public function getByRank( $rank ) {
-		$claims = new self();
-
-		/* @var Claim $claim */
-		foreach ( $this as $claim ) {
-			if ( $claim->getRank() === $rank ) {
-				$claims[] = $claim;
-			}
-		}
-
-		return $claims;
-	}
-
-	/**
-	 * Returns a new instance only containing the claims with the given ranks.
-	 *
-	 * @since 0.7.2
-	 *
-	 * @param int[] $ranks
-	 *
-	 * @return Claims
-	 */
-	public function getByRanks( array $ranks ) {
-		$ranks = array_flip( $ranks );
-		$claims = new self();
-
-		/* @var Claim $claim */
-		foreach ( $this as $claim ) {
-			if ( isset( $ranks[$claim->getRank()] ) ) {
-				$claims[] = $claim;
-			}
-		}
-
-		return $claims;
-	}
-
-	/**
-	 * Returns a new instance only containing the best claims (these are the highest
-	 * ranked claims, but never deprecated ones). This implementation ignores the properties
-	 * so you probably want to call Claims::getClaimsForProperty first or use
-	 * StatementList::getBestStatementPerProperty instead.
-	 *
-	 * @see Claims::getClaimsForProperty
-	 * @see StatementList::getBestStatementPerProperty
-	 *
-	 * @since 0.7
-	 *
-	 * @return Claims
-	 */
-	public function getBestClaims() {
-		$rank = Statement::RANK_PREFERRED;
-
-		do {
-			$claims = $this->getByRank( $rank );
-			$rank--;
-		} while ( $claims->isEmpty() && $rank > Statement::RANK_DEPRECATED );
-
-		return $claims;
-	}
-
-	/**
-	 * @see Comparable::equals
-	 *
-	 * @since 0.7.4
-	 *
-	 * @param mixed $target
-	 *
-	 * @return bool
-	 */
-	public function equals( $target ) {
-		if ( $this === $target ) {
-			return true;
-		}
-
-		if ( !( $target instanceof self )
-			|| $this->count() !== $target->count()
-		) {
-			return false;
-		}
-
-		foreach ( $this as $claim ) {
-			if ( !$target->hasExactClaim( $claim ) ) {
-				return false;
-			}
-		}
-
-		return true;
-	}
-
-	private function hasExactClaim( Claim $claim ) {
-		return $this->hasClaim( $claim )
-			&& $this->getClaimWithGuid( $claim->getGuid() )->equals( $claim );
 	}
 
 }