+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -9,6 +9,13 @@
 	* `Statement::setClaim` and `Statement::getClaim` have been removed
 	* Removed `ClaimList`
 	* Removed `ClaimListAccess`
+* Removed `Claims::getBestClaims` (you can use `StatementList::getBestStatements` instead)
+* Removed `Claims::getByRank` and `Claims::getByRanks` (you can use `StatementList::getWithRank` instead)
+* Removed `Claims::getMainSnaks` (you can use `StatementList::getMainSnaks` instead)
+* Removed `Claims::getClaimsForProperty` (you can use `StatementList::getWithPropertyId` instead)
+* Removed `Claims::getHashes`
+* Removed `Claims::getGuids`
+* Removed `Claims::equals` (and `Claims` no longer implements `Comparable`)
 
 ## Version 2.6.0 (2015-03-08)